STSMACOM-787 Export new `advancedSearchQueryToRows` helper to be used in Inventory app to reduce code duplication
