+++ lib/SearchAndSort/SearchAndSort.js
@@ -55,11 +55,17 @@ import {
 } from './nsQueryFunctions';
 import makeConnectedSource from './ConnectedSource';
 import Tags from '../Tags';
-import { NoResultsMessage, ResetButton, CollapseFilterPaneButton, ExpandFilterPaneButton } from './components';
+import {
+  NoResultsMessage,
+  ResetButton,
+  CollapseFilterPaneButton,
+  ExpandFilterPaneButton,
+} from './components';
 import { ColumnManager } from '../ColumnManager';
 import PersistedPaneset from '../PersistedPaneset';
 
 import buildUrl from './buildUrl';
+import advancedSearchQueryToRows from './advancedSearchQueryToRows';
 
 import css from './SearchAndSort.css';
 
@@ -925,7 +931,7 @@ class SearchAndSort extends React.Component {
         onSearch={this.handleAdvancedSearch}
         onCancel={() => this.setState({ isAdvancedSearchOpen: false })}
         queryBuilder={advancedSearchQueryBuilder}
-        queryToRow={this.advancedSearchQueryToRow}
+        queryToRow={(row) => advancedSearchQueryToRows(row.query)}
         hasQueryOption={false}
       >
         {({ resetRows }) => (
@@ -1164,46 +1170,6 @@ class SearchAndSort extends React.Component {
       query: searchString,
       qindex: advancedSearchIndex,
       ...extraParamsToReset,
-    });
-  };
-
-  advancedSearchQueryToRow = (row) => {
-    const formattedQuery = row.query;
-
-    if (!formattedQuery) {
-      return [];
-    }
-
-    const splitIntoRowsRegex = /(?=\sor\s|\sand\s|\snot\s)/g;
-
-    // split will return array of strings:
-    // ['keyword==test', 'or issn=123', ...]
-    const matches = formattedQuery.split(splitIntoRowsRegex).map(i => i.trim());
-
-    return matches.map((match, index) => {
-      let bool = '';
-      let query = match;
-
-      // first row doesn't have a bool operator
-      if (index !== 0) {
-        bool = match.substr(0, match.indexOf(' '));
-        query = match.substr(bool.length);
-      }
-
-      const splitIndexAndQueryRegex = /([^=]+)(exactPhrase|containsAll|startsWith)(.+)/g;
-
-      const rowParts = [...query.matchAll(splitIndexAndQueryRegex)]?.[0] || [];
-      // eslint-disable-next-line no-unused-vars
-      const [, option, _match, value] = rowParts
-        .map(i => i.trim())
-        .map(i => i.replaceAll('"', ''));
-
-      return {
-        query: value,
-        bool,
-        searchOption: option,
-        match: _match,
-      };
     });
   };
