+++ lib/SearchAndSort/index.js
@@ -1,3 +1,4 @@
 export { default } from './SearchAndSort';
 export { default as makeQueryFunction } from './makeQueryFunction';
 export { default as SearchAndSortQuery } from './SearchAndSortQuery';
+export { default as advancedSearchQueryToRows } from './advancedSearchQueryToRows';
